Week - 5/taxi_monitoring/requirements.txt
mage
tqdm
requests
joblib
pyarrow
psycopg[binary]
evidently==0.6.7
pandas
scikit-learn
jupyter